Change percent to be an integer for compatibility reasons

+++ db/migrate/20200421192921_create_discounts.rb
@@ -2,7 +2,7 @@ class CreateDiscounts < ActiveRecord::Migration[5.1]
   def change
     create_table :discounts do |t|
       t.string :name
-      t.float :percent
+      t.integer :percent
       t.integer :threshold
       t.references :merchant, foreign_key: true
